+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for specifying custom load-balancer names (@grzesiek)
 * Support specifying a fake region by environment variable (@timoreimann)
+* Add logging for Create and Update requests to the LB API (@morrislaw)
 
 ## v0.1.23 (beta) - Jan 31th 2020